fix(subject-course): renaming imports to new folder
